:zap: 3.5.1: Remove the function
